+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,8 @@
 package lotto
 
+import lotto.controller.LottoController
+import lotto.lottogenerator.RandomLottoGenerator
+
 fun main() {
     val lottoController = LottoController(RandomLottoGenerator())
     lottoController.run()
