+++ src/main/kotlin/lotto/LottoStore.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.lottogenerator.LottoGenerator
+
 class LottoStore(val lottoGenerator: LottoGenerator) {
 
     fun sellLottoTickets(purchaseAmount: LottoPurchaseAmount): List<Lotto> {
